+++ curler/__init__.py
@@ -1,3 +1,3 @@
-"""Curler Manuscript package."""
+"""Curler Paperback package."""
 
-__version__ = "0.1.0"
+__version__ = "0.2.0"